Reorganise the directory structure.
The idea here is to remove make the whole thing less confusing for users. All
source code samples are now in a chapter-specific folder and these folders have
moved to the top level.

I've also removed the 'hubbub' directory as its exact purpose is unclear. In
the future, we'll probably create a dedicated 'hubbub' repository that people
can contribute to without it having to stay in sync with the book. It could be
used to show readers more advanced techniques or the use of plugins that we
couldn't cover in the book itself.

The new 'attic' directory contains the source code from the 1st edition of the
book for those chapters we have yet to work on.